+++ src/renate/defaults.py
@@ -98,6 +98,10 @@ SER_CLS_PLASTIC_MODEL_UPDATE_PROBABILITY = 0.9
 # EWC
 EWC_LAMBDA = 0.4
 
+# LwF
+LWF_ALPHA = 1
+LWF_TEMPERATURE = 2
+
 MEMORY_SIZE = 32
